feat: added Authentication to code

+++ src/main/java/ch/hftm/boundry/BlogResource.java
@@ -12,6 +12,10 @@ import ch.hftm.control.BlogService;
 import ch.hftm.control.dto.BlogDto.NewBlogDto;
 import ch.hftm.control.dto.CommentDto.NewBlogCommentDto;
 import ch.hftm.entity.Blog;
+import io.quarkus.security.Authenticated;
+import io.quarkus.security.ForbiddenException;
+import io.quarkus.security.identity.SecurityIdentity;
+import jakarta.annotation.security.PermitAll;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.validation.Valid;
@@ -42,8 +46,12 @@ public class BlogResource {
     @Inject
     Logger logger;
 
+    @Inject
+    SecurityIdentity securityIdentity;
+
     @GET
     @Tag(name = BLOG_OVERVIEW)
+    @PermitAll
     public List<Blog> getEntries(@QueryParam("search") String searchString) {
         if (searchString != null && !searchString.isBlank())
             return this.blogService.getBlogs(searchString);
@@ -58,8 +66,9 @@ public class BlogResource {
             @APIResponse(responseCode = "201", description = "Blog created"),
             @APIResponse(responseCode = "400", description = "Invalid input")
     })
+    @Authenticated
     public Response addBlog(@Valid NewBlogDto blogDto, @Context UriInfo uriInfo) {
-        long id = this.blogService.addBlogDto(blogDto);
+        long id = this.blogService.addBlogDto(blogDto, securityIdentity.getPrincipal().getName());
         var uri = uriInfo.getAbsolutePathBuilder().path(Long.toString(id)).build();
         return Response.created(uri).build();
     }
@@ -70,6 +79,7 @@ public class BlogResource {
     @APIResponses({ @APIResponse(responseCode = "404", description = "Blog not found"),
             @APIResponse(responseCode = "200", description = "Blog found")
     })
+    @PermitAll
     public Blog getBlog(long id) {
         return this.blogService.getBlog(id)
                 .orElseThrow(() -> new NotFoundException("Blog with id " + id + " not found"));
@@ -81,10 +91,23 @@ public class BlogResource {
     @APIResponses({ @APIResponse(responseCode = "404", description = "Blog not found"),
             @APIResponse(responseCode = "204", description = "Blog deleted")
     })
+    @Authenticated
     public void deleteBlog(long id) {
         Blog blog = this.blogService.getBlog(id)
-                .orElseThrow(() -> new NotFoundException("Blog with id " + id + " not found"));
-        this.blogService.removeBlog(blog);
+            .orElseThrow(() -> new NotFoundException("Blog with id " + id + " not found"));
+
+        if(
+            securityIdentity.getRoles().contains("admin") ||
+            blog.getAuthor().equals(securityIdentity.getPrincipal().getName())
+            ) {
+            this.blogService.removeBlog(blog);
+        }
+        else {
+            throw new ForbiddenException();
+        }
+
+
+        
     }
 
     @PUT
@@ -93,6 +116,7 @@ public class BlogResource {
     @APIResponses({ @APIResponse(responseCode = "404", description = "Blog not found"),
             @APIResponse(responseCode = "201", description = "Blog updated")
     })
+    @Authenticated
     public Response updateLikedByMe(long id, @Context UriInfo uriInfo) {
         Blog blog = this.blogService.getBlog(id)
                 .orElseThrow(() -> new NotFoundException("Blog with id " + id + " not found"));
@@ -107,11 +131,12 @@ public class BlogResource {
     @Tag(name = BLOG_DETAIL)
     @Operation(description = "ID from Blog.")
     @APIResponses({
-        @APIResponse(responseCode = "500", description = "Could not create Comment"),
-        @APIResponse(responseCode = "201", description = "Comment created"),
-        @APIResponse(responseCode = "400", description = "Invalid input")
-})
-    public Response addComment(long id,@Valid NewBlogCommentDto blogCommentDto, @Context UriInfo uriInfo) {
+            @APIResponse(responseCode = "500", description = "Could not create Comment"),
+            @APIResponse(responseCode = "201", description = "Comment created"),
+            @APIResponse(responseCode = "400", description = "Invalid input")
+    })
+    @Authenticated
+    public Response addComment(long id, @Valid NewBlogCommentDto blogCommentDto, @Context UriInfo uriInfo) {
         long n_id = this.blogService.addBlogCommentDto(blogCommentDto, id);
         var uri = uriInfo.getAbsolutePathBuilder().path(Long.toString(n_id)).build();
         return Response.created(uri).build();
